agents/ai-agent-1/requirements.txt
google-adk==1.14.0
langchain==0.3.27
langchain-community==0.3.27
wikipedia==1.4.0
fastapi
uvicorn
python-dotenv
pytz